Move out weekday from airflow.contrib
